updated index.js with sotck api implementation
